requirements.txt
# Browser automation
selenium==4.18.1
webdriver-manager==4.0.1

# Scheduler
apscheduler==3.10.4

# Environment variables
python-dotenv==1.0.1
